cursor/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
pydantic